Renameing high_level module to container
