+++ cat4py/__init__.py
@@ -21,5 +21,5 @@ BITSHUFFLE = 2
 DELTA = 3
 TRUNC_PREC = 4
 
-from .high_level import Container, empty, from_buffer, from_numpy, from_file
+from .container import Container, empty, from_buffer, from_numpy, from_file
 from .version import version as __version__
